IBM System Z Cloud Provider code cleanup (part 2)
* Variables were renamed for clarity
* More error messsages were prefaced with detail for clarity
* Non-CloudProvider interface functions were moved to a helper file
